+++ src/components/base/index.ts
@@ -1,4 +1,4 @@
-export { default as Button } from './Button/Button';
+export { default as Button } from './Button';
 export { default as ChoiceButton } from './ChoiceButton';
 export { default as Input } from './Input';
 export { default as Modal } from './Modal';
